+++ lua/mini/trailspace.lua
@@ -136,6 +136,13 @@ MiniTrailspace.trim = function()
   vim.api.nvim_win_set_cursor(0, curpos)
 end
 
+--- Trim last blank lines
+MiniTrailspace.trim_last_lines = function()
+  local n_lines = vim.api.nvim_buf_line_count(0)
+  local last_nonblank = vim.fn.prevnonblank(n_lines)
+  if last_nonblank < n_lines then vim.api.nvim_buf_set_lines(0, last_nonblank, n_lines, true, {}) end
+end
+
 --- Track normal buffer
 ---
 --- Designed to be used with |autocmd|. No need to use it directly.